feat: add shared library for errors and responses
