+++ backend/auth-service/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "auth-service"
-version = "0.0.5"
+version = "0.0.6"
 edition = "2024"
 build = "build.rs"
 
@@ -25,6 +25,7 @@ tonic = "0.14.1"
 tonic-prost = "0.14.1"
 tower-http = { version = "0.6.6", features = ["cors"] }
 uuid = { version = "1.18.0", features = ["serde", "v4"] }
+shared-types = { path = "../shared-types" }
 
 [build-dependencies]
 tonic-prost-build = "0.14.1"
