+++ backend/email-service/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "email-service"
-version = "0.0.4"
+version = "0.0.5"
 edition = "2024"
 build = "build.rs"
