+++ backend/settings-service/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "settings-service"
-version = "0.0.1"
+version = "0.0.2"
 edition = "2024"
 
 [dependencies]
@@ -19,3 +19,4 @@ serde = { version = "1.0.219", features = ["derive", "rc"] }
 tokio = { version = "1.47.1", features = ["full"] }
 tower-http = { version = "0.6.6", features = ["cors"] }
 uuid = { version = "1.18.0", features = ["serde", "v4"] }
+shared-types = { path = "../shared-types" }
