+++ backend/shared-types/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "shared-types"
+version = "0.0.1"
+edition = "2024"
+
+[dependencies]
+axum = "0.8.4"
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.142"
+uuid = { version = "1.18.0", features = ["serde", "v4"] }
+chrono = { version = "0.4.41", features = ["serde"] }
+jsonwebtoken = "9.3.1"
+deadpool = "0.12.2"
+diesel = { version = "2.2.12", features = ["uuid", "chrono"] }
+diesel-async = { version = "0.6.1", features = [
+    "deadpool",
+    "pool",
+    "postgres",
+] }
+tonic = "0.14.1"
