+++ backend/shared-types/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod error;
+pub mod response;
+
+pub use error::Error;
+pub use response::*;
